requirements.txt
-r ./package_requirements.txt
pytest
pytest-django
pytest-pythonpath
mock
factory_boy

pip>=9
ipython
